+++ network/mod/ComfyNetworkSense/ComfyNetworkSense.cs
@@ -21,7 +21,7 @@ using UnityEngine;
 public sealed class ComfyNetworkSense : BaseUnityPlugin {
   public const string PluginGuid = "djcdevelopment.valheim.comfynetworksense";
   public const string PluginName = "ComfyNetworkSense";
-  public const string PluginVersion = "0.5.23";
+  public const string PluginVersion = "0.5.24";
 
   public static ComfyNetworkSense Instance { get; private set; }